src/nlp/named_entity_extractor/pyproject.toml
[tool.poetry]
name = "ner_model_training"
version = "0.1.0"
description = ""
authors = ["Attila Nagy <attila.nagy@meltwater.com>"]

[tool.poetry.dependencies]
python = "~3.6.10"
keras = "2.2.4"
tensorflow = "1.15.5"
seqeval = "1.2.2"
seaborn = "0.11.2"
scikit-learn = "0.24.1"
joblib = "1.0.1"
attrs = "20.3.0"
pyyaml = "5.4.1"
click = "7.1.2"
dotmap = "1.3.30"
falcon = "3.1.1"
gunicorn = "20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
